% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_abstract.R
\name{doc_abstract}
\alias{doc_abstract}
\title{Abstract.}
\usage{
doc_abstract(abstract, keywords)
}
\arguments{
\item{abstract}{Character string.
Abstract with \code{Jekdoc} tags.}

\item{keywords}{Character string.
Keywords separated by commas.}
}
\description{
Abstract.
}
\author{
Ivan Jacob Agaloos Pesigan
}
