% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_attr.R
\name{doc_attr}
\alias{doc_attr}
\title{Extract Asciidoctor Attribute Value Using Attribute Key.}
\usage{
doc_attr(key, attr_from_file, attr_path, attr)
}
\arguments{
\item{key}{Character vector.
Asciidoctor attribute key.}

\item{attr_from_file}{Logical.
If \code{TRUE},
attributes (\code{attr}) are extracted from a file.}

\item{attr_path}{Character string.
Path to attributes file.
Ignored if \code{attr_from_file = FALSE}.}

\item{attr}{Character vector.
Asciidoctor attributes key-value pairs.
Ignored if \code{attr_from_file = TRUE}.}
}
\description{
Extract Asciidoctor Attribute Value Using Attribute Key.
}
\author{
Ivan Jacob Agaloos Pesigan
}
