% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_attr_authorheader.R
\name{doc_attr_authorheader}
\alias{doc_attr_authorheader}
\title{Extract Author and Email from Asciidoctor Attributes for Asciidoctor Header.}
\usage{
doc_attr_authorheader(attr_from_file, attr_path, attr, format)
}
\arguments{
\item{attr_from_file}{Logical.
If \code{TRUE},
attributes (\code{attr}) are extracted from a file.}

\item{attr_path}{Character string.
Path to attributes file.
Ignored if \code{attr_from_file = FALSE}.}

\item{attr}{Character vector.
Asciidoctor attributes key-value pairs.
Ignored if \code{attr_from_file = TRUE}.}

\item{format}{Character string.
\code{"adoc"}, Asciidoctor formating.
\code{"latex"}, Latex formating.}
}
\description{
Extract Author and Email from Asciidoctor Attributes for Asciidoctor Header.
}
\author{
Ivan Jacob Agaloos Pesigan
}
