% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_attr_authorinfo.R
\name{doc_attr_authorinfo}
\alias{doc_attr_authorinfo}
\title{Extract Author Information from Asciidoctor Attributes.}
\usage{
doc_attr_authorinfo(attr_from_file, attr_path, attr, format)
}
\arguments{
\item{attr_from_file}{Logical.
If \code{TRUE},
attributes (\code{attr}) are extracted from a file.}

\item{attr_path}{Character string.
Path to attributes file.
Ignored if \code{attr_from_file = FALSE}.}

\item{attr}{Character vector.
Asciidoctor attributes key-value pairs.
Ignored if \code{attr_from_file = TRUE}.}

\item{format}{Character string.
\code{"adoc"}, Asciidoctor formating.
\code{"latex"}, Latex formating.}
}
\description{
Extract Author Information from Asciidoctor Attributes.
}
\author{
Ivan Jacob Agaloos Pesigan
}
