% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_bib.R
\name{doc_bib}
\alias{doc_bib}
\title{Bibliography.}
\usage{
doc_bib(bib_path, format)
}
\arguments{
\item{bib_path}{Character string.
Path to Biblatex file.}

\item{format}{Character string.
\code{"adoc"}, Asciidoctor formating.
\code{"latex"}, Latex formating.}
}
\description{
Bibliography.
}
\author{
Ivan Jacob Agaloos Pesigan
}
