% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_bib_latex.R
\name{doc_bib_latex}
\alias{doc_bib_latex}
\title{Bibliography.}
\usage{
doc_bib_latex(bib_path, style = "apa")
}
\arguments{
\item{bib_path}{Character string.
Path to Biblatex file.}

\item{style}{Character string.
Bibliography style.}
}
\description{
Bibliography.
}
\author{
Ivan Jacob Agaloos Pesigan
}
