% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_citation_adoc.R
\name{doc_citation_adoc}
\alias{doc_citation_adoc}
\title{Convert Jekdoc Citation Tags to Asciidoctor Bibliography Tags}
\usage{
doc_citation_adoc(from_file = TRUE, path, jdoc)
}
\arguments{
\item{from_file}{Logical.
If \code{TRUE},
Jekdoc (\code{jdoc}) is loaded from a file.}

\item{path}{Path to Jekdoc file.
Ignored if \code{from_file = FALSE}.}

\item{jdoc}{Character string.
Jekdoc input file.}
}
\description{
Convert Jekdoc Citation Tags to Asciidoctor Bibliography Tags
}
\author{
Ivan Jacob Agaloos Pesigan
}
