% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_citation_latex.R
\name{doc_citation_latex}
\alias{doc_citation_latex}
\title{Convert Jekdoc Citation Tags to Biblatex/natbib Bibliography Tags}
\usage{
doc_citation_latex(from_file = TRUE, path, jdoc, biblatex = TRUE)
}
\arguments{
\item{from_file}{Logical.
If \code{TRUE},
Jekdoc (\code{jdoc}) is loaded from a file.}

\item{path}{Path to Jekdoc file.
Ignored if \code{from_file = FALSE}.}

\item{jdoc}{Character string.
Jekdoc input file.}

\item{biblatex}{Logical.
If \code{TRUE}, convert to \code{Biblatex} tags.
If \code{FALSE}, convert to \code{natbib} tags.}
}
\description{
Convert Jekdoc Citation Tags to Biblatex/natbib Bibliography Tags
}
\author{
Ivan Jacob Agaloos Pesigan
}
