% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_doi2bib.R
\name{doc_doi2bib}
\alias{doc_doi2bib}
\title{Digital Object Identifiers to Bibtex using the Crossref API}
\usage{
doc_doi2bib(doi, par = TRUE, ncores = NULL)
}
\arguments{
\item{doi}{Character vector.
A vector of Digital Object Identifiers.}

\item{par}{Logical.
If \code{TRUE}, use multiple cores.}

\item{ncores}{Integer.
Number of cores to use if \code{par = TRUE}.
If unspecified, defaults to \code{detectCores() - 1}.}
}
\description{
Digital Object Identifiers to Bibtex using the Crossref API
}
\author{
Ivan Jacob Agaloos Pesigan
}
