% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_jdoc2adoc2html.R
\name{doc_jdoc2adoc2html}
\alias{doc_jdoc2adoc2html}
\title{Convert Jekdoc file to Asciidoc file.}
\usage{
doc_jdoc2adoc2html(
  jdoc_path,
  jdoc_abs_path,
  attr_path,
  bib_path,
  adoc = NULL,
  html = NULL,
  output_dir = getwd()
)
}
\arguments{
\item{jdoc_path}{Character string.
Jekdoc input file path
(e.g., jdoc_path = "~/sample.jdoc").}

\item{jdoc_abs_path}{Character string.
Jekdoc abstract input file path
(e.g., jdoc_abs_path = "~/abstract.jdoc").}

\item{attr_path}{Character string.
Path to attributes file.
(e.g., attr_path = "~/attributes.adoc").}

\item{bib_path}{Character string.
Path to Biblatex file.
(e.g., bib_path = "~/references.bib").}

\item{adoc}{Character string.
Asciidoc output file.
If \code{NULL},
extracts the root of \code{jdoc}.}

\item{html}{Character string.
HTML output file.
If \code{NULL},
extracts the root of \code{jdoc}.}

\item{output_dir}{Path.
Output directory.}
}
\description{
Convert Jekdoc file to Asciidoc file.
}
\author{
Ivan Jacob Agaloos Pesigan
}
