% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_knit.R
\name{doc_knit}
\alias{doc_knit}
\title{Knit Document.}
\usage{
doc_knit(input, output = NULL)
}
\arguments{
\item{input}{Character string.
Input path.}

\item{output}{Character string.
Output path.}
}
\description{
Knit Document.
}
\author{
Ivan Jacob Agaloos Pesigan
}
