% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_knit_adoc.R
\name{doc_knit_adoc}
\alias{doc_knit_adoc}
\title{Knit Jekdoc to Asciidoc.}
\usage{
doc_knit_adoc(input, output = NULL)
}
\arguments{
\item{input}{Character string.
Input path.}

\item{output}{Character string.
Output path.}
}
\description{
Converts \code{Jekdoc} \code{R} inline code and chunks to \code{Asciidoc} format
before \code{knitting}.
}
\author{
Ivan Jacob Agaloos Pesigan
}
