% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_knit_latex.R
\name{doc_knit_latex}
\alias{doc_knit_latex}
\title{Knit Jekdoc to LaTeX.}
\usage{
doc_knit_latex(input, output = NULL)
}
\arguments{
\item{input}{Character string.
Input path.}

\item{output}{Character string.
Output path.}
}
\description{
Converts \code{Jekdoc} \code{R} inline code and chunks to \code{LaTeX} format
before \code{knitting}.
}
\author{
Ivan Jacob Agaloos Pesigan
}
