% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_preamble_adoc.R
\name{doc_preamble_adoc}
\alias{doc_preamble_adoc}
\title{Asciidoc Preamble}
\usage{
doc_preamble_adoc(jdoc_abs_path, attr_path, bib_path)
}
\arguments{
\item{jdoc_abs_path}{Character string.
Jekdoc abstract input file path
(e.g., jdoc_abs_path = "~/abstract.jdoc").}

\item{attr_path}{Character string.
Path to attributes file.
(e.g., attr_path = "~/attributes.adoc").}

\item{bib_path}{Character string.
Path to Biblatex file.
(e.g., bib_path = "~/references.bib").}
}
\description{
Asciidoc Preamble
}
\author{
Ivan Jacob Agaloos Pesigan
}
