% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_preamble_latex.R
\name{doc_preamble_latex}
\alias{doc_preamble_latex}
\title{LaTeX Preamble}
\usage{
doc_preamble_latex(
  mode = "man",
  paper = "a4paper",
  packages = c("mathtools", "hyperref"),
  etc = NULL,
  jdoc_abs_path,
  attr_from_file = FALSE,
  attr_path,
  attr,
  bib_path
)
}
\arguments{
\item{mode}{Character string.
APA7 LaTeX mode.
The options are
\code{"jou"} for journal,
\code{"man"} for manuscript, and
\code{"stu"} for student paper.}

\item{paper}{Character string.
Paper size.
The options are
\code{"letter"} and
\code{"a4paper"}.}

\item{packages}{Character vector.
Additional packages to include in the preamble.}

\item{etc}{Character vector.
Extra text to include in the preamble.}

\item{jdoc_abs_path}{Path to Jekdoc abstract file.
Ignored if \code{jdoc_abs_from_file = FALSE}.}

\item{attr_from_file}{Logical.
If \code{TRUE},
attributes (\code{attr}) are extracted from a file.}

\item{attr_path}{Character string.
Path to attributes file.
Ignored if \code{attr_from_file = FALSE}.}

\item{attr}{Character vector.
Asciidoctor attributes key-value pairs.
Ignored if \code{attr_from_file = TRUE}.}

\item{bib_path}{Character string.
Path to Biblatex file.}
}
\description{
LaTeX Preamble
}
\author{
Ivan Jacob Agaloos Pesigan
}
