% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_retag.R
\name{doc_retag}
\alias{doc_retag}
\title{Replace Jekdoc Tags}
\usage{
doc_retag(tag, open, close, from_file = FALSE, path, jdoc)
}
\arguments{
\item{tag}{Character vector.
Jekdoc tags.}

\item{open}{Character vector.
Opening tag replacement.}

\item{close}{Character vector.
Closing tag replacement.}

\item{from_file}{Logical.
If \code{TRUE},
Jekdoc (\code{jdoc}) is loaded from a file.}

\item{path}{Path to Jekdoc file.
Ignored if \code{from_file = FALSE}.}

\item{jdoc}{Character string.
Jekdoc input file.}
}
\description{
Replace Jekdoc Tags
}
\author{
Ivan Jacob Agaloos Pesigan
}
