% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_retag_adoc.R
\name{doc_retag_adoc}
\alias{doc_retag_adoc}
\title{Replace Jekdoc Tags and Output Asciidoctor File}
\usage{
doc_retag_adoc(from_file = TRUE, path, jdoc)
}
\arguments{
\item{from_file}{Logical.
If \code{TRUE},
Jekdoc (\code{jdoc}) is loaded from a file.}

\item{path}{Path to Jekdoc file.
Ignored if \code{from_file = FALSE}.}

\item{jdoc}{Character string.
Jekdoc input file.}
}
\description{
Replace Jekdoc Tags and Output Asciidoctor File
}
\author{
Ivan Jacob Agaloos Pesigan
}
