% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_rm_tag.R
\name{doc_rm_tag}
\alias{doc_rm_tag}
\title{Remove Jekdoc Tags}
\usage{
doc_rm_tag(tag, jdoc)
}
\arguments{
\item{tag}{Character vector.
Jekdoc tags.}

\item{jdoc}{Character string.
Jekdoc input file.}
}
\description{
Remove Jekdoc Tags
}
\author{
Ivan Jacob Agaloos Pesigan
}
