% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_rmtag.R
\name{doc_rmtag}
\alias{doc_rmtag}
\title{Remove Jekdoc Tags}
\usage{
doc_rmtag(tag, jdoc)
}
\arguments{
\item{tag}{Character vector.
Jekdoc tags.}

\item{jdoc}{Character string.
Jekdoc input file.}
}
\description{
Remove Jekdoc Tags
}
\author{
Ivan Jacob Agaloos Pesigan
}
